requirements.txt
flask==1.1.2
flask-login==0.5.0
Flask-Mail==0.9.0
celery==5.0.1
celery[redis]
redis==3.5.3
flask-wtf==0.14.3
flask-sqlalchemy==2.4.4
pip-tools==5.3.1
black==20.8b1
pytest
pytest-cov
email_validator